Some code refactor for shells related commands
